15 implement http request and response (#17)
Implement sending http requests with cURL and getting responses for it.
Used with httpbin.org to test on real http request to check how works and correctly it sent with body, queries or headers.
Implement class for methods, protocols, host, queries, headers and even fragment to send.
There could request with queries, body or headers sent but could receive only basic response class for now.
The response contains only body as string (but could be read as bytes), status code as int and its duration it takes to be received by a server.